requirements-py31x.txt
apache-airflow
apache-airflow-providers-postgres
apache-airflow-providers-cncf-kubernetes
kubernetes
kubeflow
minio
tensorflow